+++ src/Ranking.jsx
@@ -1,46 +1,38 @@
-import { useEffect, useState } from "react";
-import axiosInstance from "../axiosInstance";
 import './css/Ranking.css';
 
-function Ranking () {
-  const [ranking, setRanking] = useState([]);
-  
-  useEffect(() => {
-    axiosInstance.get("/ranking")
-      .then(response => {
-        const sorted = response.data.sort((a, b) => b.score - a.scroe)
-        setRanking(sorted);
-      })
-      .catch(error => {
-        console.log(error);
-      })
-  }, [])
-  
+function Ranking ({gameScore}) {
+  const gameNames = [
+    { name: "LOL", label: "리그 오브 레전드" },
+    { name: "MS", label: "메이플스토리" },
+    { name: "BG", label: "배틀그라운드" },
+    { name: "SC", label: "스타크래프트" },
+  ];
   return (
     <div className="ranking-container">
-      <h2>🏆 랭킹</h2>
-      <table className="ranking-table">
-        <thead>
-          <tr>
-            <th>순위</th>
-            <th>닉네임</th>
-            <th>점수</th>
-          </tr>
-        </thead>
-        <tbody>
-          <td>123</td>
-
-          {ranking.map((user, i) => (
-            <tr key={i}>
-              <td>{i+1}</td>
-              <td>{user.nickname}</td>
-              <td>{user.score}</td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
+      <h2>🏅 랭킹</h2>
+      <p className="ranking-subtitle">지난 주 1~5위</p>
+      <div className="quiz-score-container">
+        
+          {gameNames.map((game) => (
+          <div key={game.name}>
+            <h2 className="gamename">{game.label}</h2>
+            <ul className="ranking-list">
+              {gameScore[game.name]?.map((score, i) => (
+                <li key={i} className="ranking-card">
+                  <span className="ranking-rank">{i + 1}</span>
+                  <span>{score.nickname}</span>
+                  <span className="ranking-score">{score.score}</span>
+                </li>
+              ))}
+            </ul>
+          </div>
+        ))}
+      
+      </div>
+       
     </div>
+     
   )
-}
+};
 
 export default Ranking;
